actualizado en clase

+++ src/Main.kt
@@ -1,5 +1,7 @@
+import kotlin.math.roundToInt
+
 fun Float.redondear():Float {
-    return this * 100 / 100
+    return (this * 100).toInt().toFloat() / 100
 }
 
 fun main() {
